fix(#156): stop interval when endTime is reached
setInterval is going to be trigger after first mount, and the interval
will run every x given time. To stop it, we need to clear the interval once
we reached the endTime. To do this we will need a separate effect to
keep track of the countdown and tirgger the clearInterval once the
condition is met. We will need to keep track of the interval on an
external variable outside the function as everytime the component
re-renders it will create a new interval.

It can be done more easily with setTimeout, as this will only trigger
once everytime, so we can run the effect everytime the countdown
changes, and trigger a new timeout on each re-render.

closes #156

Signed-off-by: Mauricio Robayo <rfmajo@gmail.com>